Refactor NSFW model handling and improve CI workflow
- Updated CI workflow to ensure NSFW model is downloaded only if not present.
- Added caching for Rust dependencies in CI.
- Refactored Dockerfile to download NSFW model during build if not already present.
- Moved NSFW detection logic to a separate module (nsfwcheck.rs) for better organization.
- Simplified the main application logic by delegating directory scanning to a dedicated function.
- Enhanced error handling and logging for NSFW checks.
- Updated .gitignore to exclude resources directory.
