+++ .github/workflows/ci.yml
@@ -19,7 +19,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      
+
       - name: Setup Rust
         uses: actions-rs/toolchain@v1
         with:
@@ -27,7 +27,7 @@ jobs:
           toolchain: stable
           override: true
           components: rustfmt, clippy
-      
+
       - name: Cache dependencies
         uses: actions/cache@v4
         with:
@@ -36,18 +36,20 @@ jobs:
             ~/.cargo/git
             target
           key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
-      
-      - name: Download NSFW model
+
+      - name: Ensure NSFW model is present
         run: |
           mkdir -p resources
-          curl -L -o resources/model.onnx https://github.com/Fyko/nsfw/releases/latest/download/model.onnx
-      
+          if [ ! -f resources/model.onnx ]; then
+            curl -L -o resources/model.onnx https://github.com/Fyko/nsfw/releases/latest/download/model.onnx
+          fi
+
       - name: Check formatting
         uses: actions-rs/cargo@v1
         with:
           command: fmt
           args: --all -- --check
-      
+
       - name: Run clippy
         uses: actions-rs/cargo@v1
         with:
@@ -59,14 +61,14 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      
+
       - name: Setup Rust
         uses: actions-rs/toolchain@v1
         with:
           profile: minimal
           toolchain: stable
           override: true
-      
+
       - name: Cache dependencies
         uses: actions/cache@v4
         with:
@@ -75,12 +77,14 @@ jobs:
             ~/.cargo/git
             target
           key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
-      
-      - name: Download NSFW model
+
+      - name: Ensure NSFW model is present
         run: |
           mkdir -p resources
-          curl -L -o resources/model.onnx https://github.com/Fyko/nsfw/releases/latest/download/model.onnx
-      
+          if [ ! -f resources/model.onnx ]; then
+            curl -L -o resources/model.onnx https://github.com/Fyko/nsfw/releases/latest/download/model.onnx
+          fi
+
       - name: Run tests
         uses: actions-rs/cargo@v1
         with:
@@ -104,17 +108,17 @@ jobs:
           - os: macos-latest
             artifact_name: opendirectoryscanner
             asset_name: opendirectoryscanner-macos-amd64
-    
+
     steps:
       - uses: actions/checkout@v4
-      
+
       - name: Setup Rust
         uses: actions-rs/toolchain@v1
         with:
           profile: minimal
           toolchain: stable
           override: true
-      
+
       - name: Cache dependencies
         uses: actions/cache@v4
         with:
@@ -123,26 +127,27 @@ jobs:
             ~/.cargo/git
             target
           key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
-      
-      - name: Download NSFW model
+
+      - name: Ensure NSFW model is present
         run: |
           mkdir -p resources
-          curl -L -o resources/model.onnx https://github.com/Fyko/nsfw/releases/latest/download/model.onnx
-      
+          if [ ! -f resources/model.onnx ]; then
+            curl -L -o resources/model.onnx https://github.com/Fyko/nsfw/releases/latest/download/model.onnx
+          fi
+
       - name: Build
         uses: actions-rs/cargo@v1
         with:
           command: build
           args: --release
-      
+
       - name: Upload artifact
         uses: actions/upload-artifact@v4
         with:
           name: ${{ matrix.asset_name }}
           path: target/release/${{ matrix.artifact_name }}
           if-no-files-found: error
-      
-      # Only release
+
       - name: Upload binaries to release
         if: github.event_name == 'release'
         uses: svenstaro/upload-release-action@v2
@@ -161,14 +166,14 @@ jobs:
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
-     
+
       - name: Log in to the Container registry
         uses: docker/login-action@v3
         with:
           registry: ${{ env.REGISTRY }}
           username: ${{ github.actor }}
           password: ${{ secrets.GITHUB_TOKEN }}
-     
+
       - name: Extract metadata
         id: meta
         uses: docker/metadata-action@v4
@@ -178,7 +183,7 @@ jobs:
             type=semver,pattern={{version}}
             type=ref,event=branch
             type=sha,format=short
-     
+
       - name: Build and push
         uses: docker/build-push-action@v4
         with:
